CLEAN:move strategies in dedicated folder, move classes in lib folder
