+++ lib/offer.rb
