+++ lib/offer_calculator.rb
