+++ lib/strategies/bogo_strategy.rb
