+++ lib/strategies/offer_strategy.rb
